requirements.txt
flask
transformers
torch
python-docx
sentencepiece
protobuf
tqdm
numpy
sacremoses
regex
